Rfc: 지도 생성과 위치 정보 불러오기 hook 분리
