+++ src/app/_constants/index.tsx
@@ -1,0 +1,3 @@
+export const DEFAULT_LATITUDE = 37.579293849225756;
+export const DEFAULT_LONGITUDE = 126.97798076343491;
+export const DEFAULT_ZOOM = 18;
